喵

+++ src/io/kurumi/ntt/Env.java
@@ -31,7 +31,7 @@ public class Env {
     public static String BOT_TOKEN;
 	public static String BETA_TOKEN;
 	public static String[] ALIAS;
-    public static long[] ADMINS;
+    public static int[] ADMINS;
     public static Long LOG_CHANNEL;
     public static Long TEP_CHANNEL;
 	
@@ -82,7 +82,7 @@ public class Env {
         BOT_TOKEN = conf.getStr("bot_token");
 		BETA_TOKEN = conf.getStr("beta_token");
 		ALIAS = (String[]) conf.getJSONArray("alias").toArray(String.class);
-        ADMINS = (long[]) conf.getJSONArray("admins").toArray(long.class);
+        ADMINS = (int[]) conf.getJSONArray("admins").toArray(int.class);
         LOG_CHANNEL = conf.getLong("log_channel");
         TEP_CHANNEL = conf.getLong("tep_channel");
